Adjusts axios error wrapping
The SDK API was throwing non-errors, which isn't a best practice. Axios response
json was being thrown if a request failed.

It would have saved us a lot of time today to have additional error logs available.

axios includes far more information than we typically need to debug a failed request.
Which is why it was being omitted probably.
This commit attempts to strike a balance - include the original error props,
stack trace, and a few request and response fields, while not filling the entire
logs with a single error.